worked on block-class added removal
i futherd the whole block as class scheme. I made it so that the block sort of Removes its colider, mesh and itself if its remove function is called.

also the creating the mesh and all is also now part of the block ifself.

also i noticed that i had a redundant thing going on. I was using a dictonary for the blocks themselfs and an array of all the positions of blocks so that one could iterrate though them but i realized that i could just use dictonary.keys() and get the same result. i hope i wont need to iterate though all the blocks often because of the whole class aproach.