Changed the Keybinds
i added Action 1 to 5 for various keys on the keyboard. and changed escpe as an action to "exit".

The greater idea is, that the Actions in the game have generic names and Players decide for the binigng themselfs... eventhough im 100% certain that some binds are going to be the same for every player...