continue at home
First i have to make it to where a grid while being build always saves its size, because, if a block is on the edge of a grid and it flows then it can always be considerd surrounding air.... that effectivly makes the grid smaler for volume testing.

Im Implementing the beginngins of: #4 i think a system like that would be well suited for the room discovery. And it might be able to do double duty even though it might get a little messy it might be advantagous to have it all happen in the same function....