sadly block runner isnt it. Added some more slopes...
thclfinn and me had a talk about the block runner i gotr it to be faster then what i had achived in my previous attemd but a langish buildingt would still have really bad update times of up to 10 secounds.... and that would have to become worse as more things grids are active... so we came up with an other system....

if a player builds an enclosed volume with no door or fluid port then its not like the volume is going to matter for anything... so basically, only things like doors, openable windows or fluid ports even care if a volume is enclosed. That obviously has the great benefit of only having to actually check volumes that we kinda allready know are enclosed... drastically reducing the time it should take for everything to happen.