Chage of Barrel size, first attemds to build a barrel
So this is the beginnings of fluid and parts.... and i struggle.... my first idea was to have a scene barrel that could be used in the grid but sadly that dosent work at all.... so i suppose parts will have to be build like anyother block. I hope that i can add a script to it which then is going to do what ever it is that makes that part special. That will include, telling the volume.gd of the grid that there is a volume and where its connection points are ... or for a hinge setting that up and spawaning a new grid that is atteched to the parent grid or something like that.... its a lot to consider so i kinda stall at this right now...